requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
bleach==5.0.1
blis==0.7.9
catalogue==2.0.8
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
colorama==0.4.6
confection==0.0.3
cron-descriptor==1.2.35
cryptography==39.0.0
cssselect2==0.7.0
cymem==2.0.7
distlib==0.3.6
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.1.3
django-anymail==9.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-markdownfield==0.10.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx2txt==0.8
drf-extra-fields==3.4.1
filelock==3.12.0
gunicorn==20.1.0
hashids==1.3.1
html5lib==1.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
langcodes==3.3.0
lxml==4.9.2
Markdown==3.3.4
MarkupSafe==2.1.1
msgpack==1.0.4
murmurhash==1.0.9
nltk==3.8.1
novu==1.4.0
numpy==1.24.1
oscrypto==1.3.0
packaging==22.0
pandas==1.5.2
pathy==0.10.1
paystackapi==2.1.0
pdfminer.six==20221105
pdfplumber==0.7.6
phonenumbers==8.13.3
Pillow==9.3.0
platformdirs==3.5.0
preshed==3.0.8
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==4.4.2
regex==2022.10.31
reportlab==3.6.13
requests==2.31.0
resume-parser==0.8.4
sentry-sdk==1.15.0
shortuuid==1.0.11
sib-api-v3-sdk==7.5.0
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
stemming==1.0.1
svglib==1.5.1
thinc==8.1.6
tika==2.6.0
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.13
vine==5.0.0
virtualenv==20.23.0
Wand==0.6.11
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.10